+++ json_serializable/lib/src/type_helper.dart
@@ -4,7 +4,8 @@
 
 import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/type.dart';
-import 'package:json_annotation/json_annotation.dart';
+
+import 'type_helpers/config_types.dart';
 
 /// Context information provided in calls to [TypeHelper.serialize] and
 /// [TypeHelper.deserialize].
@@ -17,11 +18,11 @@ abstract class TypeHelperContext {
 
   /// [expression] may be just the name of the field or it may an expression
   /// representing the serialization of a value.
-  Object serialize(DartType fieldType, String expression);
+  Object? serialize(DartType fieldType, String expression);
 
   /// [expression] may be just the name of the field or it may an expression
   /// representing the serialization of a value.
-  Object deserialize(DartType fieldType, String expression);
+  Object? deserialize(DartType fieldType, String expression);
 
   /// Adds [memberContent] to the set of generated, top-level members.
   void addMember(String memberContent);
@@ -30,7 +31,7 @@ abstract class TypeHelperContext {
 /// Extended context information with includes configuration values
 /// corresponding to `JsonSerializableGenerator` settings.
 abstract class TypeHelperContextWithConfig extends TypeHelperContext {
-  JsonSerializable get config;
+  ClassConfig get config;
 }
 
 abstract class TypeHelper<T extends TypeHelperContext> {
@@ -51,7 +52,7 @@ abstract class TypeHelper<T extends TypeHelperContext> {
   /// String serialize(DartType targetType, String expression) =>
   ///   "$expression.id";
   /// ```.
-  Object serialize(DartType targetType, String expression, T context);
+  Object? serialize(DartType targetType, String expression, T context);
 
   /// Returns Dart code that deserializes an [expression] representing a JSON
   /// literal to into [targetType].
@@ -76,7 +77,7 @@ abstract class TypeHelper<T extends TypeHelperContext> {
   /// String deserialize(DartType targetType, String expression) =>
   ///   "new ${targetType.name}.fromInt($expression)";
   /// ```.
-  Object deserialize(
+  Object? deserialize(
     DartType targetType,
     String expression,
     T context,