+++ json_serializable/lib/src/type_helpers/generic_factory_helper.dart
@@ -6,26 +6,33 @@ import 'package:analyzer/dart/element/type.dart';
 
 import '../lambda_result.dart';
 import '../type_helper.dart';
+import '../utils.dart';
 
 class GenericFactoryHelper extends TypeHelper<TypeHelperContextWithConfig> {
   const GenericFactoryHelper();
 
   @override
-  Object serialize(
+  Object? serialize(
     DartType targetType,
     String expression,
     TypeHelperContextWithConfig context,
   ) {
     if (context.config.genericArgumentFactories &&
         targetType is TypeParameterType) {
-      return LambdaResult(expression, toJsonForType(targetType));
+      final toJsonFunc = toJsonForType(targetType);
+      if (targetType.isNullableType) {
+        context.addMember(_toJsonHelper);
+        return '$_toJsonHelperName($expression, $toJsonFunc)';
+      }
+
+      return LambdaResult(expression, toJsonFunc);
     }
 
     return null;
   }
 
   @override
-  Object deserialize(
+  Object? deserialize(
     DartType targetType,
     String expression,
     TypeHelperContextWithConfig context,
@@ -33,12 +40,39 @@ class GenericFactoryHelper extends TypeHelper<TypeHelperContextWithConfig> {
   ) {
     if (context.config.genericArgumentFactories &&
         targetType is TypeParameterType) {
-      return LambdaResult(expression, fromJsonForType(targetType));
+      final fromJsonFunc = fromJsonForType(targetType);
+
+      if (targetType.isNullableType) {
+        context.addMember(_fromJsonHelper);
+        return '$_fromJsonHelperName($expression, $fromJsonFunc)';
+      }
+
+      return LambdaResult(expression, fromJsonFunc);
     }
 
     return null;
   }
 }
+
+const _fromJsonHelperName = r'_$nullableGenericFromJson';
+
+const _fromJsonHelper = '''
+T? $_fromJsonHelperName<T>(
+  Object? input,
+  T Function(Object? json) fromJson,
+) =>
+    input == null ? null : fromJson(input);
+''';
+
+const _toJsonHelperName = r'_$nullableGenericToJson';
+
+const _toJsonHelper = '''
+Object? $_toJsonHelperName<T>(
+  T? input,
+  Object? Function(T value) toJson,
+) =>
+    input == null ? null : toJson(input);
+''';
 
 String toJsonForType(TypeParameterType type) =>
     toJsonForName(type.getDisplayString(withNullability: false));