+++ json_serializable/test/generic_files/generic_argument_factories.dart
@@ -2,8 +2,6 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// @dart=2.12
-
 import 'package:json_annotation/json_annotation.dart';
 
 part 'generic_argument_factories.g.dart';
@@ -30,8 +28,8 @@ class GenericClassWithHelpers<T, S> {
       _$GenericClassWithHelpersFromJson(json, fromJsonT, fromJsonS);
 
   Map<String, dynamic> toJson(
-    Object Function(T value) toJsonT,
-    Object Function(S value) toJsonS,
+    Object? Function(T value) toJsonT,
+    Object? Function(S value) toJsonS,
   ) =>
       _$GenericClassWithHelpersToJson(this, toJsonT, toJsonS);
 }
@@ -42,7 +40,10 @@ class ConcreteClass {
 
   final GenericClassWithHelpers<double, BigInt> value2;
 
-  ConcreteClass(this.value, this.value2);
+  // Regression scenario for google/json_serializable.dart#803
+  final GenericClassWithHelpers<double?, BigInt?> value3;
+
+  ConcreteClass(this.value, this.value2, this.value3);
 
   factory ConcreteClass.fromJson(Map<String, dynamic> json) =>
       _$ConcreteClassFromJson(json);